+++ packages/fe-fpm-writer/templates/common/EventHandler.ts
@@ -1,0 +1,5 @@
+import MessageToast from 'sap/m/MessageToast';
+
+export function <%- (typeof eventHandlerFnName !== 'undefined' && eventHandlerFnName) || 'onPress' %>() {
+    MessageToast.show("Custom handler invoked.");
+}