+++ packages/fe-fpm-writer/test/test-input/basic-ts/webapp/Component.ts
@@ -1,0 +1,12 @@
+import AppComponent from "sap/fe/core/AppComponent";
+
+/**
+ * @namespace v4travel
+ */
+export default class Component extends AppComponent {
+
+	public static metadata = {
+		manifest: "json"
+	};
+
+}