+++ examples/simple-generator/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @sap-ux/generator-simple-fe
 
+## 0.2.18
+
+### Patch Changes
+
+-   Updated dependencies [bc4cb3a]
+    -   @sap-ux/btp-utils@0.11.0
+    -   @sap-ux/logger@0.3.0
+    -   @sap-ux/axios-extension@0.9.5
+
+## 0.2.17
+
+### Patch Changes
+
+-   Updated dependencies [d351f81]
+    -   @sap-ux/fiori-elements-writer@0.5.0
+    -   @sap-ux/fiori-freestyle-writer@0.13.0
+
 ## 0.2.16
 
 ### Patch Changes